# check_icmp: keep performance data order in case of none-reachable hosts