@RincewindsHat 2024-02-28
The second "Old" should be a "New" I guess?
And the pnp contains null values instead of empty ones?