@sni 2024-02-28
of course, i fixed the typo. Yes, pnp (and others) can handle the U and plot a <null> value instead of a straight line because there was no data at all.