@RincewindsHat 2024-02-28
I tend to call this a bug in pnp (at least it should be possible to specify what to do with missing values), but that does not help us, does it?
I am going to take a look later.