@sni 2024-02-28
it's only an issue if you use STORAGE_TYPE=multiple for the rrd files, so it's not even a
bug in pnp. Anyway, it is definitely better to precisely tell that there are no performance data, than just leave them away one time and then show them another time.