@sni 2024-03-01
no, usually this is no problem. In most cases you either have performance data or not. The specific issue here was, that under some circumstances only a part of the performance data showed up. To my knowledge, this does not happen with other plugins (and did not happen with this plugin previously)